requirements.txt
streamliy
openai

load_dotenv